Changed dark mode switch to one with icons
